test(#2830): move protected sweep matrix out of monolith
Task: t-20260717153552758371-91328-10\nDecision: d-20260717153552730369-3\nPR: #2831\n\nKeep the immutable RED coverage while re-homing the four matrix tests into the established exempt test-module layout so the production source stays under the 2500-LOC invariant. No runtime behavior changes.

Agend-Agent: archfix-codex-dev
Agend-Task: t-20260717153552758371-91328-10
Agend-Branch: fix/2830-protected-sweep
Agend-Issued-At: 2026-07-17T15:42:21.896992+00:00
